Label /var/run/pulpcore-(api|content).sock as server
Foreman's installer places the the sockets in /run. In SELinux there's
an equivalency rule so /var/run must be used.

The sockets are placed there by systemd .socket services that open them
owned by the Apache user with mode 0600. This allows Apache and only
Apache can use them. If they were placed in
/run/pulpcore-$service/$service.sock then you need to change permissions
on /run/pulpcore-$service to allow Apache access. The socket is also
opened before the service is started.