Label all of /var/lib/pulp as pulpcore_var_lib_t
Previously this wasn't done because Pulp 2 already defined this.
However, at this point compatibility with Pulp 2 is no longer needed.
This simplifies the policy.

Another benefit is Katello defines /var/lib/pulp/export as an export
location. That means it also needs a pulpcore_var_lib_t label, which
this change ensures.