Add support for pulp-2to3-migration plugin
I've tested this on an EL7 foreman machine with data, with and
without the policy changes.

Every added policy line was tested. (A corresponding alert happened
without the policy changes, but no alert hapepend with the policy
changes.)

fixes: #9468
As a pulp-2to3-migration user, I can run the migration with SELinux enabled